worker/requirements.txt

fastapi
uvicorn[standard]
qdrant-client
python-dotenv
requests
pymongo
playwright==1.40.0
fake-useragent
firecrawl-py
beautifulsoup4
gunicorn
google-generativeai
crewai
langchain
langchain_google_genai
numpy
pydantic
cohere
celery
redis
